+++ src/customConfigTests/src/test/kotlin/org/vaccineimpact/orderlyweb/customConfigTests/CustomConfigTests.kt
@@ -1,4 +1,4 @@
-package org.vaccineimpact.orderlyweb.customConfigTests.customconfig_tests
+package org.vaccineimpact.orderlyweb.customConfigTests
 
 import khttp.responses.Response
 import org.assertj.core.api.Assertions
